fix(billing): widen storage checkpoint totals to BigInt
lastReportedTotal/pendingReportedTotal accumulate MiB-hours over a billing
month (ADR-027: period total = Σ megabytes). Int32 caps at ~2.8 TiB held for
a full month before overflow, which a large enterprise org can exceed and
would fail the Prisma/Postgres write. Change both columns to BIGINT / Prisma
BigInt while the table is still empty, and carry bigint through the service
interface and unit tests. Per-hour megabytes stays Int (no realistic
single-hour overflow); consecutiveFailures stays Int.

+++ langwatch/prisma/migrations/20260619120000_add_storage_billing_tables/migration.sql
@@ -26,8 +26,8 @@ CREATE TABLE "StorageBillingCheckpoint" (
     "id" TEXT NOT NULL,
     "organizationId" TEXT NOT NULL,
     "billingMonth" TEXT NOT NULL,
-    "lastReportedTotal" INTEGER NOT NULL DEFAULT 0,
-    "pendingReportedTotal" INTEGER,
+    "lastReportedTotal" BIGINT NOT NULL DEFAULT 0,
+    "pendingReportedTotal" BIGINT,
     "consecutiveFailures" INTEGER NOT NULL DEFAULT 0,
     "updatedAt" TIMESTAMP(3) NOT NULL,
